pyproject.toml
[tool.poetry]
name = "mini smart home"
version = "0.1.0"
description = ""
authors = ["Eugene <vint.evg@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.12.0"
pydantic-settings = "^2.4.0"
paho-mqtt = "^2.1.0"
rq = "^1.16.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
bandit = "^1.7.9"
flake8 = "^7.1.1"
pytest = "^8.3.2"
